@@ -1,10 +1,10 @@
 name: animated_headline
 description: A simple animated headline for your app. This package is a Flutter widget.
-version: 1.1.1
+version: 1.2.0
 homepage: https://github.com/eppeque/animated_headline
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
